requirements.txt
Flask
requests
python-dotenv
Flask-WTF
pytest